spike(core): extract KlaviyoPushCore extension-safe module
