+++ KlaviyoPushCore.podspec
@@ -1,9 +1,10 @@
 Pod::Spec.new do |s|
-  s.name             = "KlaviyoCore"
+  s.name             = "KlaviyoPushCore"
   s.version          = "5.3.0"
-  s.summary          = "Core functionalities for the Klaviyo SDK"
+  s.summary          = "Extension-safe push notification primitives for the Klaviyo SDK"
   s.description      = <<-DESC
-                        Core functionalities and utilities for the Klaviyo SDK.
+                        Extension-safe shared types for push notification handling in the Klaviyo SDK.
+                        Safe to use in both main app targets and Notification Service Extension targets.
                        DESC
   s.homepage         = "https://github.com/klaviyo/klaviyo-swift-sdk"
   s.license          = { :type => "MIT", :file => "LICENSE" }
@@ -11,8 +12,5 @@ Pod::Spec.new do |s|
   s.source           = { :git => "https://github.com/klaviyo/klaviyo-swift-sdk.git", :tag => s.version.to_s }
   s.swift_version    = '5.7'
   s.platform         = :ios, '13.0'
-  s.source_files     = 'Sources/KlaviyoCore/**/*.swift'
-  s.pod_target_xcconfig = { 'OTHER_SWIFT_FLAGS' => '-package-name KlaviyoCore' }
-  s.dependency       'AnyCodable-FlightSchool'
-  s.dependency       'KlaviyoPushCore', '~> 5.3.0'
+  s.source_files     = 'Sources/KlaviyoPushCore/**/*.swift'
 end
