+++ Sources/KlaviyoPushCore/Logger+Ext.swift
@@ -1,0 +1,22 @@
+//
+//  Logger+Ext.swift
+//
+
+import OSLog
+
+@available(iOS 14.0, *)
+extension Logger {
+    private static var subsystem = "com.klaviyo.klaviyo-swift-sdk.klaviyoPushCore"
+
+    init(category: String) {
+        self.init(subsystem: Self.subsystem, category: category)
+    }
+}
+
+// MARK: - Loggers
+
+@available(iOS 14.0, *)
+extension Logger {
+    /// Logger for notification category management
+    static let notifications = Logger(category: "Notifications")
+}
