+++ Sources/KlaviyoPushCore/UNNotificationContent+Klaviyo.swift
@@ -1,0 +1,30 @@
+//
+//  UNNotificationContent+Klaviyo.swift
+//
+
+import Foundation
+import UserNotifications
+
+extension Dictionary where Key == AnyHashable {
+    /// Determines if a notification payload originated from Klaviyo.
+    ///
+    /// A notification is considered a Klaviyo notification if it contains
+    /// a "body" dictionary with a "_k" key in its userInfo.
+    ///
+    /// - Returns: `true` if the notification is from Klaviyo, `false` otherwise
+    public func isKlaviyoNotification() -> Bool {
+        guard let properties = self as? [String: Any],
+              let body = properties["body"] as? [String: Any],
+              body["_k"] != nil else {
+            return false
+        }
+        return true
+    }
+}
+
+extension UNNotificationContent {
+    /// Determines if this notification content originated from Klaviyo.
+    public var isKlaviyoNotification: Bool {
+        userInfo.isKlaviyoNotification()
+    }
+}
